[SECURITY] main(): use separate input and output FILE * structures
According to "man fopen":

    Reads  and  writes  may be intermixed on read/write streams in any
    order.  Note that ANSI C requires that a file positioning function
    intervene between output and input, unless an input operation
    encounters end-of-file.

Using fseek() does not work correctly on network sockets, so it does not
satisfy this condition.

In some cases, mixing fread and fwrite on network streams resulted in
write operations not going through. This caused the program to hang without
returning any output and made a DoS attack possible.

The solution is to use separate FILE * objects for input and output, as
suggested by:

https://www.masterraghu.com/subjects/np/introduction/unix_network_programming_v1.3/ch14lev1sec8.html

A partial workaround for the security issue is to use the -t command
line parameter (timeout) with short values, like -t2 .

Copyright (C) 2021 Hostersi, being sent with permission from my
employer.