main(): do not mix operations on FILE * and raw file descriptor - part 1
As is written in "man fgets":
       It  is  not  advisable to mix calls to input functions from the
       stdio library with low-level calls to read(2) for the file
       descriptor associated with the input stream; the results will be
       undefined and very probably not what you want.

This program mixes fgets() with poll() on the file descriptor. This can cause
issues if there is data in the C library's buffers.

This patch uses read() in place of fgets() when reading from the network socket.

Copyright (C) 2021 Hostersi, being sent with permission from my
employer.